+++ test/controller.js
@@ -43,6 +43,14 @@ describe('Controller', function(){
       });
     });
 
+    it('should respond to connected', function(){
+      var controller = fakeController()
+      assert.equal(controller.connected(), false);
+      controller.connect()
+      controller.on('ready', function() {
+        assert.equal(controller.connected(), true);
+      });
+    });
   });
 
   describe('#disconnect', function() {